Refactor Streams to use interfaces
This is required to allow proxying of the `StreamProvider` and enables more flexibility.
