+++ aff4/aff4-core/src/main/kotlin/com/github/nava2/aff4/streams/image_stream/Aff4ImageStreamSourceProvider.kt
@@ -1,129 +1,12 @@
 package com.github.nava2.aff4.streams.image_stream
 
-import com.github.nava2.aff4.container.ContainerDataFileSystemProvider
-import com.github.nava2.aff4.io.SourceProvider
-import com.github.nava2.aff4.io.buffer
-import com.github.nava2.aff4.io.concatLazily
-import com.github.nava2.aff4.io.sourceProvider
-import com.github.nava2.aff4.io.use
-import com.github.nava2.aff4.model.Aff4Model
 import com.github.nava2.aff4.model.Aff4StreamSourceProvider
 import com.github.nava2.aff4.model.VerifiableStreamProvider
-import com.github.nava2.aff4.model.VerifiableStreamProvider.Result.FailedHash
-import com.github.nava2.aff4.model.rdf.Aff4Arn
 import com.github.nava2.aff4.model.rdf.ImageStream
-import com.github.nava2.aff4.streams.computeLinearHashes
-import com.github.nava2.aff4.streams.hashingSink
-import com.google.inject.assistedinject.Assisted
-import com.google.inject.assistedinject.AssistedInject
-import okio.Source
-import okio.Timeout
 import java.io.Closeable
 
-class Aff4ImageStreamSourceProvider @AssistedInject internal constructor(
-  aff4ImageBeviesFactory: Aff4ImageBevies.Factory,
-  private val containerDataFileSystemProvider: ContainerDataFileSystemProvider,
-  @Assisted val imageStream: ImageStream,
-) : VerifiableStreamProvider, Aff4StreamSourceProvider, Closeable {
-
-  override val arn: Aff4Arn = imageStream.arn
-  override val size: Long = imageStream.size
-
-  private val aff4ImageBevies: Aff4ImageBevies = aff4ImageBeviesFactory.create(imageStream)
-  private val bevyCount = imageStream.bevyCount
-
-  @Volatile
-  private var verificationResult: VerifiableStreamProvider.Result? = null
-
-  @Volatile
-  private var closed: Boolean = false
-
-  override fun source(position: Long, timeout: Timeout): Source {
-    check(!closed) { "closed" }
-    return Aff4ImageStreamSource(
-      aff4ImageBevies = aff4ImageBevies,
-      imageStream = imageStream,
-      position = position,
-      timeout = timeout,
-    )
-  }
-
-  override fun close() {
-    if (closed) return
-    closed = true
-
-    aff4ImageBevies.close()
-  }
-
-  override fun verify(aff4Model: Aff4Model, timeout: Timeout): VerifiableStreamProvider.Result {
-    check(!closed) { "closed" }
-
-    val previousResult = verificationResult
-    if (previousResult != null) {
-      return previousResult
-    }
-
-    return synchronized(this) {
-      val doubleChecked = verificationResult
-      if (doubleChecked != null) return doubleChecked
-
-      timeout.throwIfReached()
-
-      val failedHashes = mutableListOf<FailedHash>()
-      failedHashes += verifyLinearHashes(timeout)
-      failedHashes += verifyBlockHashes(aff4Model, timeout)
-
-      val result = VerifiableStreamProvider.Result.fromFailedHashes(failedHashes)
-      verificationResult = result
-      result
-    }
-  }
-
-  override fun toString(): String {
-    return "Aff4ImageStream(${imageStream.arn}, closed=$closed)"
-  }
-
-  private fun verifyLinearHashes(timeout: Timeout): Sequence<FailedHash> = sequence {
-    val calculatedLinearHashes = use(timeout = timeout) { s ->
-      s.computeLinearHashes(imageStream.linearHashes.map { it.hashType })
-    }
-
-    val linearHashesByHashType = imageStream.linearHashes.associateBy { it.hashType }
-    for ((hashType, actualHash) in calculatedLinearHashes) {
-      val expectedHash = linearHashesByHashType.getValue(hashType)
-      if (expectedHash.value != actualHash) {
-        yield(FailedHash(imageStream, "Linear", expectedHash))
-      }
-    }
-  }
-
-  private fun verifyBlockHashes(aff4Model: Aff4Model, timeout: Timeout): Sequence<FailedHash> = sequence {
-    val blockHashes = imageStream.queryBlockHashes(aff4Model)
-
-    val blockHashSourceProviders = (0 until bevyCount).asSequence()
-      .map { aff4ImageBevies.getOrLoadBevy(it).bevy }
-      .fold(blockHashes.associateWith { mutableListOf<SourceProvider<Source>>() }) { acc, bevy ->
-        for ((hash, blockHashPath) in bevy.blockHashes) {
-          val sources = acc.entries.single { it.key.forHashType == hash }.value
-          sources += containerDataFileSystemProvider.get(imageStream).sourceProvider(blockHashPath)
-        }
-
-        acc
-      }
-
-    for ((blockHash, sourceProviders) in blockHashSourceProviders) {
-      val actualHash = sourceProviders.concatLazily().buffer().use(timeout = timeout) { source ->
-        blockHash.hash.hashType.hashingSink().use { hashSink ->
-          source.readAll(hashSink)
-          hashSink.hash
-        }
-      }
-
-      if (blockHash.hash.value != actualHash) {
-        yield(FailedHash(imageStream, "BlockHash ${blockHash.forHashType}", blockHash.hash))
-      }
-    }
-  }
+interface Aff4ImageStreamSourceProvider : VerifiableStreamProvider, Aff4StreamSourceProvider, Closeable {
+  val imageStream: ImageStream
 
   interface Loader : Aff4StreamSourceProvider.Loader<ImageStream, Aff4ImageStreamSourceProvider>
 }
