+++ aff4/aff4-core/src/main/kotlin/com/github/nava2/aff4/streams/map_stream/Aff4MapStreamModule.kt
@@ -3,9 +3,11 @@ package com.github.nava2.aff4.streams.map_stream
 import com.github.nava2.aff4.model.Aff4StreamOpener
 import com.github.nava2.aff4.model.rdf.MapStream
 import com.github.nava2.aff4.streams.AbstractAff4StreamModule
-import com.github.nava2.guice.assistedFactoryModule
+import com.github.nava2.guice.build
+import com.github.nava2.guice.implement
 import com.github.nava2.guice.key
 import com.github.nava2.guice.typeLiteral
+import com.google.inject.assistedinject.FactoryModuleBuilder
 
 internal object Aff4MapStreamModule : AbstractAff4StreamModule<MapStream, Aff4MapStreamSourceProvider>(
   configTypeLiteral = typeLiteral(),
@@ -14,6 +16,10 @@ internal object Aff4MapStreamModule : AbstractAff4StreamModule<MapStream, Aff4Ma
   override fun configureModule() {
     requireBinding<Aff4StreamOpener>()
 
-    install(assistedFactoryModule<Aff4MapStreamSourceProvider.Loader>())
+    install(
+      FactoryModuleBuilder()
+        .implement<Aff4MapStreamSourceProvider, RealAff4MapStreamSourceProvider>()
+        .build<Aff4MapStreamSourceProvider.Loader>()
+    )
   }
 }
